+++ src/components/progressBar.js
@@ -1,6 +1,6 @@
-import { getNewsContent } from "./api/api.js";
+import { getNewsContent } from "../api/api.js";
 import { drawListView } from "./listNews.js";
-import { setDisplay } from "./util/utils.js";
+import { setDisplay } from "../util/utils.js";
 
 const CATEGORY_NUM = 7;
 const PROGRESS_TIME = 2000;
