+++ src/components/randomGrid.js
@@ -1,6 +1,6 @@
-import { UNSUB_BTN_IMG, GRID_SUB_BTN_IMG } from "./path.js";
-import { getPressObj } from "./api/api.js";
-import { subscribeState } from "./store/subscribeState.js";
+import { UNSUB_BTN_IMG, GRID_SUB_BTN_IMG } from "../path.js";
+import { getPressObj } from "../api/api.js";
+import { subscribeState } from "../store/subscribeState.js";
 
 const PRESS_NUM_IN_GRID = 24;
 const TOTAL_PRESS_NUM = 96;
